+++ HISTORY.md
@@ -7,6 +7,7 @@
 * Update Color to Level 4 (2025-04-24)
 * Update Values and Units to Level 4 (WD 2024-03-12)
 * Update Display Level 3 to CR 2023-03-30
+* Add support for Ruby Level 1 (WD 2022-12-31)
 
 ## css-sanitizer 5.5.0 (2025-01-27)
 * Ensure <-token and identifiers are always separated as a security