+++ HISTORY.md
@@ -6,6 +6,7 @@
 	- properties: aspect-ratio, contain-intrinsic-* (size, width, height, block-size, inline-size), min-intrinsic-size;
 * Update Color to Level 4 (2025-04-24)
 * Update Values and Units to Level 4 (WD 2024-03-12)
+* Update Display Level 3 to CR 2023-03-30
 
 ## css-sanitizer 5.5.0 (2025-01-27)
 * Ensure <-token and identifiers are always separated as a security